% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\docType{data}
\name{OAuth}
\alias{OAuth}
\title{OAuth}
\value{
an \link{OAuth} object
}
\description{
OAuth
}
\details{
\strong{Methods}
\describe{
\item{\code{url()}}{
get the url
}
\item{\code{as_string()}}{
get the url
}
\item{\code{is_valid()}}{
check if it's valid
}
\item{\code{signed_attributes()}}{
get the signed attributes
}
}
}
\examples{
\dontrun{
(x <- OAuth$new("get", "https://api.twitter.com/1/statuses/friends.json", list()))
x
x$url()
x$as_string()
x$signed_attributes()
x$is_valid()

(x <- OAuth$new("get", "https://api.twitter.com/1/statuses/friends.json?stuff=4", 
  list(), list(consumer_secret = "hello", token_secret = "world")))
x$options
x$.__enclos_env__$private$attributes()
x$.__enclos_env__$private$normalized_attributes()
x$.__enclos_env__$private$signature()
x$.__enclos_env__$private$hmac_sha1_signature()
x$.__enclos_env__$private$secret()
x$.__enclos_env__$private$signature_base()
x$.__enclos_env__$private$normalized_params()
x$.__enclos_env__$private$signature_params()
x$.__enclos_env__$private$url_params()
x$.__enclos_env__$private$rsa_sha1_signature()
x$.__enclos_env__$private$private_key()
}
}
\keyword{datasets}
